requirements.txt
diffusers==0.23.0
transformers==4.34.1
gradio==3.50.2
--extra-index-url https://download.pytorch.org/whl/cu121;
torch==2.1.0
fastapi==0.104.1
uvicorn==0.24.0.post1
Pillow==10.1.0
accelerate==0.24.0
compel==2.0.2
controlnet-aux==0.0.7
peft==0.6.0
xformers; sys_platform != 'darwin' or platform_machine != 'arm64'